fastapi_test/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
asyncpg
greenlet